Use template.py in workflow.py and issue.py
