+++ exasol/toolbox/tools/issue.py
@@ -1,30 +1,12 @@
-import difflib
-import io
-from contextlib import ExitStack
-from pathlib import Path
-from typing import (
-    Any,
-    Mapping,
-    Union,
-)
-
-import importlib_resources as resources
 import typer
+from pathlib import Path
 from rich.columns import Columns
-from rich.console import Console
-from rich.syntax import Syntax
-
-stdout = Console()
-stderr = Console(stderr=True)
+import exasol.toolbox.tools.template as template
 
 CLI = typer.Typer()
-
-from exasol.toolbox.tools import template
-
-
-def _issues() -> Mapping[str, Any]:
-    pkg = "exasol.toolbox.templates.github.ISSUE_TEMPLATE"
-    return template._templates(pkg)
+pkg = "exasol.toolbox.templates.github.ISSUE_TEMPLATE"
+template_type = "issue"
+lexer = "markdown"
 
 
 @CLI.command(name="list")
@@ -34,30 +16,16 @@ def list_issues(
     )
 ) -> None:
     """List all available issues."""
+    template.list_templates(columns, pkg) 
 
-    class List:
-        def __init__(self, items: Any):
-            self.items = items
-
-        def __rich__(self) -> str:
-            return "\n".join(self.items)
-
-    output = List(_issues()) if not columns else Columns(_issues(), expand=True)
-    stdout.print(output)
 
 
 @CLI.command(name="show")
 def show_issue(
-    issue: str = typer.Argument(..., help="Issue which shall be shown."),
+    issue: str
 ) -> None:
     """Shows a specific issue."""
-    issues = _issues()
-    if issue not in issues:
-        stdout.print(f"Unknown issue <{issue}>.", style="red")
-        raise typer.Exit(code=1)
-
-    issue = issues[issue]
-    stdout.print(Syntax.from_path(path=issue, encoding="utf-8", lexer="markdown"))  # type: ignore
+    template.show_templates(issue, pkg, template_type, lexer)
 
 
 @CLI.command(name="diff")
@@ -69,58 +37,14 @@ def diff_issue(
     ),
 ) -> None:
     """Diff a specific issue against the installed one."""
-    issues = _issues()
-    if issue not in issues:
-        stdout.print(f"Unknown issue <{issue}>.", style="red")
-        raise typer.Exit(code=1)
-
-    # Use Any type to enable reuse of the variable/binding name
-    old: Any = dest / f"{issue}.md"
-    new: Any = Path(_issues()[issue])
-    with ExitStack() as stack:
-        old = stack.enter_context(
-            open(old, encoding="utf-8") if old.exists() else io.StringIO("")
-        )
-        new = stack.enter_context(open(new, encoding="utf-8"))
-        old = old.read().split("\n")
-        new = new.read().split("\n")
-
-    diff = difflib.unified_diff(old, new, fromfile="old", tofile="new")
-    stdout.print(Syntax("\n".join(diff), "diff"))
-
-
-def _install_issue(
-    src: Union[str, Path], dest: Union[str, Path], exists_ok: bool = False
-) -> None:
-    src, dest = Path(src), Path(dest)
-
-    if dest.exists() and not exists_ok:
-        raise FileExistsError("Issue already exists")
-
-    with ExitStack() as stack:
-        input_file = stack.enter_context(open(src, "rb"))
-        output_file = stack.enter_context(open(dest, "wb"))
-        output_file.write(input_file.read())
-
-
-def _select_issues(issue: str) -> Mapping[str, Any]:
-    issues = _issues()
-    if issue != "all" and issue not in issues:
-        raise ValueError(f"Issue <{issue}> is unknown")
-    issues = (
-        issues
-        if issue == "all"
-        else {name: path for name, path in issues.items() if issue == name}
-    )
-    return issues
+    template.diff_template(issue, dest, pkg, template_type)
 
 
 @CLI.command(name="install")
 def install_issue(
     issue: str = typer.Argument("all", help="name of the issue to install."),
     dest: Path = typer.Argument(
-        Path("./.github/ISSUE_TEMPLATE"),
-        help="target directory to install the issue to.",
+        Path("./.github/ISSUE_TEMPLATE"), help="target directory to install the issue to."
     ),
 ) -> None:
     """
@@ -128,60 +52,21 @@ def install_issue(
 
     Attention: If there is an existing issue with the same name it will be overwritten!
     """
-    if not dest.exists():
-        dest.mkdir()
-
-    try:
-        issues = _select_issues(issue)
-    except Exception as ex:
-        stderr.print(f"[red]{ex}[/red]")
-        raise typer.Exit(-1)
-
-    for name, path in issues.items():
-        destination = dest / f"{name}.md"
-        _install_issue(path, destination, exists_ok=True)
-        stderr.print(f"Installed {name} in {destination}")
+    template.install_template(issue, dest, pkg, template_type)
 
 
 @CLI.command(name="update")
 def update_issue(
     issue: str = typer.Argument("all", help="name of the issue to install."),
     dest: Path = typer.Argument(
-        Path("./.github/ISSUE_TEMPLATE"),
-        help="target directory to install the issue to.",
+        Path("./.github/ISSUE_TEMPLATE"), help="target directory to install the issue to."
     ),
     confirm: bool = typer.Option(
         False, help="Automatically confirm overwritting exsisting issue(s)"
     ),
 ) -> None:
     """Similar to install but checks for existing issues and shows diff"""
-    if not dest.exists():
-        dest.mkdir()
-
-    try:
-        issues = _select_issues(issue)
-    except Exception as ex:
-        stderr.print(f"[red]{ex}[/red]")
-        raise typer.Exit(-1)
-
-    if confirm:
-        install_issue(issue, dest)
-        raise typer.Exit(0)
-
-    for name, path in issues.items():
-        destination = dest / f"{name}.md"
-        try:
-            _install_issue(path, destination, exists_ok=False)
-            stderr.print(f"Updated {name} in {destination}")
-        except FileExistsError:
-            show_diff = typer.confirm(f"issue <{name}> already exists, show diff?")
-            if show_diff:
-                diff_issue(name, dest)
-
-            overwrite = typer.confirm(f"Overwrite existing issue?")
-            if overwrite:
-                _install_issue(path, destination, exists_ok=True)
-                stderr.print(f"Updated {name} in {destination}")
+    template.update_template(issue, dest, confirm, pkg, template_type)
 
 
 if __name__ == "__main__":
